requirements.txt
streamlit
sentence-transformers
chromadb
Pillow
torch